[REFACTOR] 파일 구조 변경 및 사용하지 않는 props 제거
- atoms/SideMenuList/SideMenuList => modules/SideMenu/SideMenuList
- sideMenuList사용할 때  props.type 제거
